More endianness fixes

+++ src/m_swap.h
@@ -38,9 +38,9 @@
 // WAD files are stored little endian.
 #ifdef WORDS_BIGENDIAN
 short	SwapSHORT(short);
-long	SwapLONG(long);
+int	SwapLONG(int);
 #define SHORT(x)	((short)SwapSHORT((unsigned short) (x)))
-#define LONG(x)         ((long)SwapLONG((unsigned long) (x)))
+#define LONG(x)         ((int)SwapLONG((unsigned int) (x)))
 #else
 #define SHORT(x)	(x)
 #define LONG(x)         (x)
@@ -53,6 +53,9 @@ long	SwapLONG(long);
 //-----------------------------------------------------------------------------
 //
 // $Log$
+// Revision 1.7  2006/01/24 01:46:08  fraggle
+// More endianness fixes
+//
 // Revision 1.6  2006/01/23 21:56:18  fraggle
 // Include the config header so that endianness is dealt with correctly
 //
